refactor: split up table, improve search logic
